docs(istio): add k3s installation instructions for ambient mode
Adds a note and a specific command for installing Istio ambient mode on a k3s cluster using the flannel CNI. This is required as the platform needs to be specified for the CNI directories to be configured correctly.

Also updates the main documentation link to point directly to the ambient section.

Signed-off-by: owan <owan.io1992@gmail.com>